[PA] Pack frequently accessed items closer together in the thread cache.
In the thread cache, the first buckets are hotter, since small
allocations are more frequent than large ones. However the stats, which
are accessed for each allocation, are grouped towards the end of the
structure. While all these members do not fit in a single cacheline, and
there is no real false sharing risk as the thread cache size class is
large enough, it is still better to group the hot data together.

However, on 64 bit architectures, this makes the first 2 buckets plus
the per-allocation metadata fit into 2 cachelines, so may increase cache
locality for the smaller allocations.

Bug: 1207307
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3113971
Commit-Queue: Benoit L <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#915123}

+++ base/allocator/partition_allocator/partition_stats.h
@@ -33,8 +33,8 @@ struct ThreadCacheStats {
   uint64_t batch_fill_count;  // Number of central allocator requests.
 
   // Memory cost:
-  uint64_t bucket_total_memory;
-  uint64_t metadata_overhead;
+  uint32_t bucket_total_memory;
+  uint32_t metadata_overhead;
 
 #if defined(PA_THREAD_CACHE_ALLOC_STATS)
   uint64_t allocs_per_bucket_[kNumBuckets + 1];
